Completed Chatbot and ticket
