--- front-end/src/components/Agent/Agent.css
@@ -1,65 +1,0 @@
-.agent {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-  }
-  
-  .agent__messages {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    width: 100%;
-    min-height: 400px;
-    max-height: 400px;
-    overflow-y: auto;
-    border: 1px solid #ccc;
-    padding: 10px;
-    margin-bottom: 20px;
-  }
-  
-  .agent__message {
-    max-width: 80%;
-    padding: 5px 10px;
-    border-radius: 10px;
-    margin-bottom: 10px;
-  }
-  
-  .agent__message.user {
-    background-color: #f1f1f1;
-  }
-  
-  .agent__message.agent {
-    background-color: #2196f3;
-    color: white;
-    align-self: flex-end;
-  }
-  
-  .agent__input {
-    display: flex;
-    width: 100%;
-  }
-  
-  .agent__input input {
-    flex-grow: 1;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px 0 0 5px;
-  }
-  
-  .agent__input button {
-    padding: 10px;
-    border: none;
-    background-color: #2196f3;
-    color: white;
-    border-radius: 0 5px 5px 0;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  .agent__input button:hover {
-    background-color: #0b7dda;
-  }
-
