+++ front-end/src/components/Agent/Agent.module.css
@@ -1,0 +1,102 @@
+.container {
+  font-family: 'Roboto', sans-serif;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 1rem;
+}
+
+.title {
+  font-size: 2rem;
+  margin-bottom: 1rem;
+}
+
+.subtitle {
+  font-size: 1.5rem;
+  margin-bottom: 1rem;
+}
+
+.request-list {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+.request-item {
+  margin-bottom: 0.5rem;
+  display: flex;
+  align-items: center;
+}
+
+.request-button {
+  margin-left: 0.5rem;
+}
+
+.chat-container {
+  border: 1px solid #e5e5e5;
+  border-radius: 5px;
+  padding: 1rem;
+  background-color: #f9f9f9;
+}
+
+.message-list {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  max-height: 300px;
+  overflow-y: auto;
+}
+
+.message-item {
+  margin-bottom: 0.5rem;
+}
+
+.right-align {
+  text-align: right;
+}
+
+.left-align {
+  text-align: left;
+}
+
+.form-container {
+  display: flex;
+  margin-bottom: 1rem;
+}
+
+.input-message {
+  flex: 1;
+  padding: 0.5rem;
+  border-radius: 5px;
+  border: 1px solid #e5e5e5;
+  outline: none;
+}
+
+.submit-button {
+  margin-left: 0.5rem;
+}
+
+.end-chat-button {
+  background-color: #f44336;
+  color: white;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem 1rem;
+  border-radius: 5px;
+}
+
+button {
+  background-color: #3f51b5;
+  color: white;
+  border: none;
+  cursor: pointer;
+  padding: 0.5rem 1rem;
+  border-radius: 5px;
+}
+
+button:hover {
+  background-color: #283593;
+}
+
+.end-chat-button:hover {
+  background-color: #c62828;
+}
